Added Junit test cases Added completed implementation of user lock service

+++ implementation/domain/logic/jbb-security-impl/src/main/java/org/jbb/security/impl/lock/properties/UserLockProperties.java
@@ -19,7 +19,7 @@ import org.jbb.lib.properties.ModuleProperties;
 public interface UserLockProperties extends ModuleProperties { //NOSONAR
 
     @Key("user.lock.maximum.sign.in.attempt")
-    long userSignInMaximumAttempt();
+    int userSignInMaximumAttempt();
 
     @Key("user.lock.time.period.in.minutes")
     long userSignInLockTimePeriod();
